maskrcnn_coco/pyproject.toml
[project]
name = "maskrcnn-coco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cjm-pandas-utils>=0.0.3",
    "cjm-pil-utils>=0.0.9",
    "cjm-psl-utils>=0.0.4",
    "cjm-pytorch-utils>=0.0.8",
    "cjm-torchvision-tfms>=0.0.26",
    "distinctipy>=1.3.4",
    "fiftyone>=1.10.0",
    "gym>=0.26.2",
    "gym-anytrading>=2.0.0",
    "gymnasium>=1.2.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pip>=25.3",
    "pycocotools>=2.0.10",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "stable-baselines3>=2.7.0",
    "tabulate>=0.9.0",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchtnt>=0.2.4",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.225",
    "wandb>=0.23.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
